Refactoring call api requests & Separate constants by files according to their purpose
