+++ src/actions/auth.js
@@ -1,7 +1,5 @@
-import fetch from 'isomorphic-fetch';
-import * as types from '../constants';
-
-const api = 'localhost:8000';
+import callApi from '../utils/call-api';
+import * as types from '../constants/auth';
 
 export function signup(username, password) {
   return dispatch => {
@@ -9,24 +7,12 @@ export function signup(username, password) {
       type: types.SIGNUP_REQUEST
     });
 
-    return fetch(`http://${api}/v1/signup`, {
-      method: 'POST',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        username,
-        password
-      })
-    })
-      .then(response => response.json())
-      .then(json => {
-        if (!json.success) {
-          throw new Error(json.message);
-        }
-        return json;
-      })
+    return callApi(
+      'signup',
+      undefined,
+      { method: 'POST' },
+      { username, password }
+    )
       .then(json => {
         if (!json.token) {
           throw new Error('Token has not been provided!');
@@ -55,24 +41,12 @@ export function login(username, password) {
       type: types.LOGIN_REQUEST
     });
 
-    return fetch(`http://${api}/v1/login`, {
-      method: 'POST',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        username,
-        password
-      })
-    })
-      .then(response => response.json())
-      .then(json => {
-        if (!json.success) {
-          throw new Error(json.message);
-        }
-        return json;
-      })
+    return callApi(
+      '/login',
+      undefined,
+      { method: 'POST' },
+      { username, password }
+    )
       .then(json => {
         if (!json.token) {
           throw new Error('Token has not been provided!');
@@ -101,20 +75,7 @@ export function logout() {
       type: types.LOGOUT_REQUEST
     });
 
-    return fetch(`http://${api}/v1/logout`, {
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body: ''
-    })
-      .then(response => response.json())
-      .then(json => {
-        if (!json.success) {
-          throw new Error(json.message);
-        }
-        return json;
-      })
+    return callApi('/logout')
       .then(json =>
         dispatch({
           type: types.LOGOUT_SUCCESS
@@ -137,36 +98,25 @@ export function receiveAuth() {
 
     const { token } = getState().auth;
 
-    return fetch(`http://${api}/v1/users/me`, {
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${token}`
-      }
-    })
-      // simulate delay from api
-      .then(
-        response =>
-          new Promise(resolve => setTimeout(() => resolve(response), 500))
-      )
-      .then(response => response.json())
-      .then(json => {
-        if (!json.success) {
-          throw new Error(json.message);
-        }
-        return json;
-      })
-      .then(json =>
-        dispatch({
-          type: types.RECEIVE_AUTH_SUCCESS,
-          payload: json
+    return (
+      callApi('users/me', token)
+        // simulate delay from api
+        .then(
+          response =>
+            new Promise(resolve => setTimeout(() => resolve(response), 500))
+        )
+        .then(json =>
+          dispatch({
+            type: types.RECEIVE_AUTH_SUCCESS,
+            payload: json
+          })
+        )
+        .catch(reason => {
+          dispatch({
+            type: types.RECEIVE_AUTH_FAILURE,
+            payload: reason
+          });
         })
-      )
-      .catch(reason => {
-        dispatch({
-          type: types.RECEIVE_AUTH_FAILURE,
-          payload: reason
-        });
-      });
+    );
   };
 }
