+++ src/constants/index.js
@@ -1,17 +1,2 @@
-export const SIGNUP_REQUEST = Symbol('SIGNUP_REQUEST');
-export const SIGNUP_SUCCESS = Symbol('SIGNUP_SUCCESS');
-export const SIGNUP_FAILURE = Symbol('SIGNUP_FAILURE');
-
-export const LOGIN_REQUEST = Symbol('LOGIN_REQUEST');
-export const LOGIN_SUCCESS = Symbol('LOGIN_SUCCESS');
-export const LOGIN_FAILURE = Symbol('LOGIN_FAILURE');
-
-export const LOGOUT_REQUEST = Symbol('LOGOUT_REQUEST');
-export const LOGOUT_SUCCESS = Symbol('LOGOUT_SUCCESS');
-export const LOGOUT_FAILURE = Symbol('LOGOUT_FAILURE');
-
-export const RECEIVE_AUTH_REQUEST = Symbol('RECEIVE_AUTH_REQUEST');
-export const RECEIVE_AUTH_SUCCESS = Symbol('RECEIVE_AUTH_SUCCESS');
-export const RECEIVE_AUTH_FAILURE = Symbol('RECEIVE_AUTH_FAILURE');
-
-export const REDIRECT = Symbol('REDIRECT');
+export * from './auth';
+export * from './services';
